'Handle NULL input properly and handle larger number for binary conv'

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -1,6 +1,6 @@
 #include "main.h"
 
-void binary(unsigned int);
+void binary(unsigned long);
 /**
  * print_binary - Print the binary convertion of a decimal
  * @n: decimal number
@@ -21,7 +21,7 @@ void print_binary(unsigned long int n)
  * binary - print binary
  * @n: number
  */
-void binary(unsigned int n)
+void binary(unsigned long n)
 {
 	int bit;
